update moments for normalize bug (#444)

+++ nvtabular/ops/moments.py
@@ -125,7 +125,7 @@ def _chunkwise_moments(df):
         df2[col] = df[col].astype("float64").pow(2)
     vals = {
         "df-count": df.count().to_frame().transpose(),
-        "df-sum": df.sum().to_frame().transpose(),
+        "df-sum": df.sum().astype("float64").to_frame().transpose(),
         "df2-sum": df2.sum().to_frame().transpose(),
     }
     # NOTE: Perhaps we should convert to pandas here
